fix: changed int to Long for the projectId.

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/domain/model/commands/CreateMaterialCommand.java
@@ -1,7 +1,7 @@
 package com.construtech.buildsphere.platform.resourceManagement.domain.model.commands;
 
 public record CreateMaterialCommand(String materialName, String materialDescription, String receptionDate,
-                                    int amount, double totalCost, String status, int projectId) {
+                                    int amount, double totalCost, String status, Long projectId) {
     public CreateMaterialCommand {
         if (projectId <= 0) {
             throw new IllegalArgumentException("Project ID must be greater than 0");
